requirements.txt
fastapi
uvicorn
python-multipart
python-jose
passlib[bcrypt]

pytest
httpx
ruff
